requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
requests==2.28.2
Werkzeug==2.2.3
SQLAlchemy==2.0.7